move activation quantizers from qat to trainable infrastructure for pytorch
