+++ model_compression_toolkit/qat/pytorch/quantizer/ste_rounding/uniform_ste.py
@@ -20,26 +20,25 @@ from torch import Tensor
 from model_compression_toolkit.constants import RANGE_MAX, RANGE_MIN
 from model_compression_toolkit.trainable_infrastructure.common.constants import FQ_MIN, FQ_MAX
 
-from model_compression_toolkit.qat import TrainingMethod
 from model_compression_toolkit.target_platform_capabilities.target_platform import QuantizationMethod
 from mct_quantizers import QuantizationTarget, PytorchQuantizationWrapper
-from model_compression_toolkit import constants as C
 
-from model_compression_toolkit.qat.pytorch.quantizer.base_pytorch_qat_quantizer import BasePytorchQATTrainableQuantizer
+from model_compression_toolkit.qat.pytorch.quantizer.base_pytorch_qat_weight_quantizer import BasePytorchQATWeightTrainableQuantizer
 from mct_quantizers import mark_quantizer
 from model_compression_toolkit.core.pytorch.utils import to_torch_tensor
-from model_compression_toolkit.qat.pytorch.quantizer.quantizer_utils import uniform_quantizer
+from model_compression_toolkit.trainable_infrastructure import TrainingMethod
+from model_compression_toolkit.trainable_infrastructure.pytorch.quantizer_utils import uniform_quantizer
 from mct_quantizers.pytorch.quantizers import \
-    WeightsUniformInferableQuantizer, ActivationUniformInferableQuantizer
+    WeightsUniformInferableQuantizer
 from model_compression_toolkit.trainable_infrastructure.common.trainable_quantizer_config import \
-    TrainableQuantizerWeightsConfig, TrainableQuantizerActivationConfig
+    TrainableQuantizerWeightsConfig
 from model_compression_toolkit.trainable_infrastructure.common.base_trainable_quantizer import VariableGroup
 
 
 @mark_quantizer(quantization_target=QuantizationTarget.Weights,
                 quantization_method=[QuantizationMethod.UNIFORM],
                 identifier=TrainingMethod.STE)
-class STEUniformWeightQATQuantizer(BasePytorchQATTrainableQuantizer):
+class STEUniformWeightQATQuantizer(BasePytorchQATWeightTrainableQuantizer):
     """
     Trainable constrained quantizer to quantize a layer inputs.
     """
@@ -66,7 +65,6 @@ class STEUniformWeightQATQuantizer(BasePytorchQATTrainableQuantizer):
                               [-1]) if self.quantization_config.weights_per_channel_threshold else float(
             self.min_values)
 
-
     def initialize_quantization(self,
                                 tensor_shape: torch.Size,
                                 name: str,
@@ -87,7 +85,6 @@ class STEUniformWeightQATQuantizer(BasePytorchQATTrainableQuantizer):
         # Save the quantizer parameters for later calculations
         self.add_quantizer_variable(FQ_MIN, layer.get_parameter(name+"_"+FQ_MIN), VariableGroup.QPARAMS)
         self.add_quantizer_variable(FQ_MAX, layer.get_parameter(name+"_"+FQ_MAX), VariableGroup.QPARAMS)
-
 
     def __call__(self,
                  inputs: nn.Parameter,
@@ -117,79 +114,3 @@ class STEUniformWeightQATQuantizer(BasePytorchQATTrainableQuantizer):
                                                 max_range=_max.tolist(),
                                                 per_channel=self.quantization_config.weights_per_channel_threshold,
                                                 channel_axis=self.quantization_config.weights_channels_axis)
-
-
-@mark_quantizer(quantization_target=QuantizationTarget.Activation,
-                quantization_method=[QuantizationMethod.UNIFORM],
-                identifier=TrainingMethod.STE)
-class STEUniformActivationQATQuantizer(BasePytorchQATTrainableQuantizer):
-    """
-    Trainable constrained quantizer to quantize a layer activations.
-    """
-
-    def __init__(self, quantization_config: TrainableQuantizerActivationConfig):
-        """
-        Initialize a STEUniformActivationQATQuantizer object with parameters to use
-        for uniform quantization.
-
-        Args:
-            quantization_config: trainable quantizer config class
-        """
-        super().__init__(quantization_config)
-
-        np_min_range = quantization_config.activation_quantization_params[C.RANGE_MIN]
-        np_max_range = quantization_config.activation_quantization_params[C.RANGE_MAX]
-        self.min_range_tensor = torch.Tensor([np_min_range])
-        self.max_range_tensor = torch.Tensor([np_max_range])
-        self.num_bits = quantization_config.activation_n_bits
-
-    def initialize_quantization(self,
-                                tensor_shape: torch.Size,
-                                name: str,
-                                layer: PytorchQuantizationWrapper):
-        """
-        Add quantizer parameters to the quantizer parameters dictionary
-
-        Args:
-            tensor_shape: tensor shape of the quantized tensor.
-            name: Tensor name.
-            layer: Layer to quantize.
-        """
-        layer.register_parameter(name+"_"+FQ_MIN, nn.Parameter(to_torch_tensor(self.min_range_tensor), requires_grad=True))
-        layer.register_parameter(name+"_"+FQ_MAX, nn.Parameter(to_torch_tensor(self.max_range_tensor), requires_grad=True))
-
-        # Save the quantizer parameters for later calculations
-        self.add_quantizer_variable(FQ_MIN, layer.get_parameter(name+"_"+FQ_MIN), VariableGroup.QPARAMS)
-        self.add_quantizer_variable(FQ_MAX, layer.get_parameter(name+"_"+FQ_MAX), VariableGroup.QPARAMS)
-
-    def __call__(self,
-                 inputs: torch.Tensor,
-                 training: bool = True) -> torch.Tensor:
-        """
-        Quantize a tensor.
-        Args:
-            inputs: Input tensor to quantize.
-            training: Whether the graph is in training mode.
-
-        Returns:
-            The quantized tensor.
-        """
-
-        _min = self.get_quantizer_variable(FQ_MIN)
-        _max = self.get_quantizer_variable(FQ_MAX)
-        q_tensor = uniform_quantizer(inputs, _min, _max, self.num_bits)
-        return q_tensor
-
-    def convert2inferable(self) -> ActivationUniformInferableQuantizer:
-        """
-        Convert quantizer to inferable quantizer.
-
-        Returns:
-            A pytorch inferable quanizer object.
-        """
-        _min = self.get_quantizer_variable(FQ_MIN).cpu().detach().numpy()
-        _max = self.get_quantizer_variable(FQ_MAX).cpu().detach().numpy()
-
-        return ActivationUniformInferableQuantizer(num_bits=self.num_bits,
-                                                   min_range=_min.tolist(),
-                                                   max_range=_max.tolist())
