+++ model_compression_toolkit/trainable_infrastructure/common/training_method.py
@@ -1,4 +1,4 @@
-# Copyright 2022 Sony Semiconductor Israel, Inc. All rights reserved.
+# Copyright 2024 Sony Semiconductor Israel, Inc. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,7 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-from model_compression_toolkit.qat.common.qat_config import QATConfig
+from enum import Enum
 
-from model_compression_toolkit.qat.keras.quantization_facade import keras_quantization_aware_training_init_experimental, keras_quantization_aware_training_finalize_experimental
-from model_compression_toolkit.qat.pytorch.quantization_facade import pytorch_quantization_aware_training_init_experimental, pytorch_quantization_aware_training_finalize_experimental
+
+class TrainingMethod(Enum):
+    STE = "STE"    # Standard straight-through estimator
+    DQA = "DQA"    # DNN Quantization with Attention. Includes a smooth quantization introduced by DQA method
+    LSQ = "LSQ"    # Learned Step size Quantization. https://arxiv.org/pdf/1902.08153.pdf
