+++ model_compression_toolkit/trainable_infrastructure/pytorch/base_pytorch_quantizer.py
@@ -12,31 +12,21 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-from typing import Union, List
+from typing import List
+
+from abc import ABC
 
 from model_compression_toolkit.logger import Logger
 from model_compression_toolkit.verify_packages import FOUND_TORCH
 from model_compression_toolkit.trainable_infrastructure.common.base_trainable_quantizer import VariableGroup
 from model_compression_toolkit.trainable_infrastructure.common.base_trainable_quantizer import BaseTrainableQuantizer, VAR, GROUP
-from model_compression_toolkit.trainable_infrastructure import TrainableQuantizerWeightsConfig, \
-    TrainableQuantizerActivationConfig
 
 
 if FOUND_TORCH:
 
     import torch
 
-    class BasePytorchTrainableQuantizer(BaseTrainableQuantizer):
-        def __init__(self,
-                     quantization_config: Union[TrainableQuantizerWeightsConfig, TrainableQuantizerActivationConfig]):
-            """
-            This class is a base Pytorch quantizer which validates the provided quantization config and defines an
-            abstract function which any quantizer needs to implement.
-
-            Args:
-                quantization_config: quantizer config class contains all the information about the quantizer configuration.
-            """
-            super().__init__(quantization_config)
+    class BasePytorchTrainableQuantizer(BaseTrainableQuantizer, ABC):
 
         def get_trainable_variables(self, group: VariableGroup) -> List[torch.Tensor]:
             """
