+++ tests/pytorch_tests/trainable_infrastructure_tests/trainable_pytorch/test_pytorch_base_quantizer.py
@@ -12,6 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
+from typing import List, Any
+
+from model_compression_toolkit.trainable_infrastructure.common.base_trainable_quantizer import VariableGroup
 from model_compression_toolkit.trainable_infrastructure.common.trainable_quantizer_config import \
     TrainableQuantizerWeightsConfig, TrainableQuantizerActivationConfig
 from model_compression_toolkit.trainable_infrastructure.pytorch.base_pytorch_quantizer import \
@@ -78,8 +81,18 @@ class TestPytorchBaseActivationQuantizer(BasePytorchInfrastructureTest):
 
 
 class _TestQuantizer(BasePytorchTrainableQuantizer):
+
     def __init__(self, quantization_config: TrainableQuantizerWeightsConfig):
         super().__init__(quantization_config)
+
+    def get_trainable_variables(self, group: VariableGroup) -> List[Any]:
+        pass
+
+    def initialize_quantization(self, tensor_shape, name: str, layer):
+        pass
+
+    def __call__(self, input2quantize, training: bool):
+        pass
 
 
 class TestPytorchQuantizerWithoutMarkDecorator(BasePytorchInfrastructureTest):
